xml/parsing/manakai/peref-1.dat

#data
<!DOCTYPE a [
%a;
]><a/>
#errors
1;1;no XML decl;;;s
2;1;pe not declared;;a;;m
2;1;stop processing;;;i
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a [
%abc
]><a/>
#errors
1;1;no XML decl;;;s
3;0;no refc;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a [
%abc<!ENTITY abc "yy">
]><a x="&abc;"/>
#errors
1;1;no XML decl;;;s
2;5;no refc;;;m
#document
| <!DOCTYPE a>
| <{} a>
|   x="yy"

#data
<!DOCTYPE a [
  <!ENTITY % def "">
%abc%def;
]><a x="&abc;"/>
#errors
1;1;no XML decl;;;s
3;5;no refc;;;m
4;9;entity not declared;;abc;;m
#document
| <!DOCTYPE a>
| <{} a>
|   x="&abc;"

#data
<!DOCTYPE a [
  <!ENTITY % def "">
%abc!?aa;
]><a/>
#errors
1;1;no XML decl;;;s
3;1;stop processing;;;i
3;1;pe not declared;;abc!?aa;;m
#document
| <!DOCTYPE a>
| <{} a>
